fixed usage of wrong file, created turn class
