+++ lib/turn.rb
@@ -1,15 +1,3 @@
-require 'minitest/autorun'
-require 'minitest/pride'
-require './lib/card'
-require './lib/turn'
-
-class TurnTest < Minitest::Test
-
-  def test_it_exists
-    card_1 = Card.new("What is the capital of Alaska?", "Juneau", :Geography)
-    turn = Turn.new("Junaeu", card_1)
-
-    assert_instance_of Turn, turn
-  end
+class Turn
 
 end
