+++ test/turn_test.rb
@@ -1,0 +1,15 @@
+require 'minitest/autorun'
+require 'minitest/pride'
+require './lib/card'
+require './lib/turn'
+
+class TurnTest < Minitest::Test
+
+  def test_it_exists
+    card_1 = Card.new("What is the capital of Alaska?", "Juneau", :Geography)
+    turn = Turn.new("Junaeu", card_1)
+
+    assert_instance_of Turn, turn
+  end
+
+end
